api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
pydantic==2.10.0
python-dotenv==1.0.0
mangum==0.17.0
